refactored create_and_write_dict_list to write_dict_list
it now contains the logic for determining if the file already exists and simply adds new rows to it. if the file doesn't exist it creates the new file, the header, and then adds the rows to it. also we are no longer using the write_dict.keys() to determine the fieldname parameter. It was not matching things up, and only worked in a specific order. For this reason I created a new list transaction_write_list and it holds the column names that match up with incoming dicts to be written

+++ csv_utils.py
@@ -3,6 +3,20 @@
 from csv import DictReader, DictWriter
 import os
 
+transaction_write_list = [
+    "gmail_message_id",
+    "gmail_thread_id",
+    "authorized_time",
+    "amount",
+    "vendor",
+    "account",
+    "condensed_message",
+    "Subject",
+    "From",
+    "To",
+    "Date"
+]
+
 def read_csv(filepath):
     data = []
     with open(filepath, 'r', newline='\n') as csv_file:
@@ -20,50 +34,17 @@ def delete_file(filepath):
         return "{} did not exist in local directory".format(filepath)
     return "removed {} from local directory".format(filepath)
 
-def create_csv_and_write_dict_list(filepath, write_dict_list):
-    with open(filepath, 'w', newline='\n') as csv_file:
-        writer = DictWriter(csv_file, fieldnames=write_dict_list[0].keys())
+def write_dict_list(filepath, write_dict_list):
+    if not os.path.exists(filepath):
+        with open(filepath, 'w', newline='\n') as csv_file:
+            writer = DictWriter(csv_file, fieldnames=transaction_write_list)
 
-        writer.writeheader()
+            writer.writeheader()
 
-        writer.writerows(write_dict_list)
+            writer.writerows(write_dict_list)
+    else:
+        with open(filepath, 'a', newline='\n') as csv_file:
+            writer = DictWriter(csv_file, fieldnames=transaction_write_list)
+            writer.writerows(write_dict_list)
 
     return "file written to {}".format(filepath)
-
-def write_csv(filepath, columns, data):
-    with open(filepath, 'w', newline='\n') as csv_file:
-        writer = DictWriter(csv_file, fieldnames=columns)
-        
-        writer.writeheader()
-
-        writer.writerows(data)
-    
-    return "file written to {}".format(filepath)
-
-def delete_row_from_csv(filepath, row_number):
-    current_data = read_csv(filepath)
-    print(current_data)
-    # recreate current_data - row_number
-    overwritten_data = []
-    for i, order in enumerate(current_data):
-        if(i != row_number):
-            overwritten_data.append(order)
-        # if(i == row_number):
-        #     print("i: {}\nrow_number: {}\nrow: {}".format(i, row_number, order))
-    write_csv(filepath, overwritten_data[0].keys(), overwritten_data)
-
-
-def write_row_to_csv(filepath, row):
-    with open(filepath, 'a', newline='\n') as csv_file:
-        DictWriter(csv_file, fieldnames=row.keys()).writerow(row)
-
-    return "row added to {}".format(filepath)
-
-def delete_row(row_number):
-    print("about to delete: {}\nfrom: {}".format(row_number, 'reports/pending_orders/buy_orders.csv'))
-    delete_row_from_csv('reports/pending_orders/buy_orders.csv', row_number)
-
-def delete_rows(row_numbers):
-    print("about to delete: {}".format(row_numbers))
-    for row_number in row_numbers:
-        delete_row(row_number)
